chore(editor-state): relocating files and splitting viewport helpers
