+++ packages/editor/packages/editor-state/src/impureHelpers/viewport/index.ts
@@ -1,0 +1,4 @@
+export { calculateBorderLineCoordinates } from './calculateBorderLineCoordinates';
+export { move } from './move';
+export { resize } from './resize';
+export { snapToGrid } from './snapToGrid';
